backend/requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
openai==1.109.1
pinecone-client==3.1.0
numpy>=1.26.0
python-dotenv==1.0.0
requests==2.32.3
